Refactor patient views into multiple views
