+++ components/textContainers.js/InputField.js
@@ -6,7 +6,7 @@ export function InputField({ name, label, type, value, onChange }) {
       </label>
       <input
         id="first_name"
-        class="text-sm rounded-lg  block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 "
+        class="text-sm rounded-lg block w-full  bg-green-200 "
         name={name}
         className="input"
         type={type}
