+++ pages/record.js
@@ -3,13 +3,11 @@ import axios from "axios";
 import _ from "lodash";
 import Modal from "react-modal";
 import moment from "moment";
-import {
-  ConsultationsTable,
-  ConsultationsView,
-  VitalsView,
-  VisitPrescriptionsTable,
-  PatientView,
-} from "../components/views/patient";
+import { ConsultationsView } from "@/components/views/Consultations/ConsultationsView";
+import { VitalsView } from "@/components/views/Vitals/VitalsView";
+import { ConsultationsTable } from "@/components/views/Consultations/ConsultationsTable";
+import { VisitPrescriptionsTable } from "@/components/views/Prescriptions/VisitPrescriptionsTable";
+import { PatientView } from "../components/views/patient";
 import { API_URL, CLOUDINARY_URL } from "../utils/constants";
 import withAuth from "../utils/auth";
 import Router from "next/router";
